Move `git_off_t` to `git_object_size_t`
`git_off_t` is problematic: first, it's used to represent the size of
a blob, but that's represented in git as an `unsigned long`, meaning
that we've lost a bit (on 64-bit machines).  Second, `git_off_t` is a
poor name, suggesting that it's an offset, not a size (which makes
sense, given that it's a signed type).

Migrate `git_off_t` to `git_object_size_t`, an unsigned 64-bit type.